# Refactor project structure to lisp letter case
