--- src/deep-root-angular/Tests/Backend/node-bin/preinstall.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
-
-[ -e "./node_modules/deepify" ] && rm -f "./node_modules/deepify";
-
-[ -e "./node_modules/AsyncConfig" ] && rm -f "./node_modules/AsyncConfig";
-
-[ -e "./node_modules/Scheduler" ] && rm -f "./node_modules/Scheduler";
-
-
-exit 0
