+++ src/deep-root-angular/backend/resources.json
