+++ src/deep-root-angular/frontend/bootstrap.js
