+++ src/deep-root-angular/frontend/img/favicon.ico
