+++ src/deep-root-angular/frontend/js/app/directives/index.js
