+++ src/deep-root-angular/frontend/js/app/directives/updateTitle.js
