+++ src/deep-root-angular/frontend/js/app/index.js
