+++ src/deep-root-angular/frontend/js/app/module/index.js
