+++ src/deep-root-angular/frontend/js/app/module/ng-config.js
