+++ src/deep-root-angular/frontend/js/app/module/ng-module.js
