+++ src/deep-root-angular/frontend/js/lib/angular-ui-router.js
