+++ src/deep-root-angular/frontend/js/lib/angular.js
