+++ src/deep-root-angular/frontend/js/package.json
