+++ src/deep-root-angular/tests/backend/node-bin/install.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+npm link chai &&\
+npm link aws-sdk &&\
+npm link node-dir &&\
+npm link deepify &&\
+ln -s ../../../Backend/src/async-config ./node_modules &&\
+ln -s ../../../Backend/src/scheduler ./node_modules
