+++ src/deep-root-angular/tests/backend/node-bin/preinstall.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+[ -e "./node_modules/deepify" ] && rm -f "./node_modules/deepify";
+
+[ -e "./node_modules/async-config" ] && rm -f "./node_modules/async-config";
+
+[ -e "./node_modules/scheduler" ] && rm -f "./node_modules/scheduler";
+
+
+exit 0
