+++ src/deep-root-angular/tests/backend/test/scheduler/rule/bootstrap.spec.js
@@ -3,11 +3,11 @@
 'use strict';
 
 import chai from 'chai';
-import bootstrap from '../../../node_modules/AsyncConfig/Dump/bootstrap';
+import bootstrap from '../../../node_modules/scheduler/rule/bootstrap';
 
 // @todo: Add more advanced tests
 suite('Bootstraps', () => {
-  test(' bootstrap exists in deep-system-async-config-dump module', () => {
+  test(' bootstrap exists in deep-system-scheduler-rule module', () => {
     chai.expect(bootstrap).to.be.an('object');
   });
 });
