+++ src/deep-root-angular/tests/frontend/config.karma.js
