+++ src/deep-root-angular/tests/frontend/config.test.js
