+++ src/deep-root-angular/tests/frontend/mock/lib/DeepFramework.js
