+++ src/deep-root-angular/tests/frontend/package.json
